refactor: extract config value editor to reusable module
